userspace: split address-book table building out of policies.go
Move the address-book building concern verbatim from policies.go into a new
policies_addrbook.go: the content-ID collision plumbing
(addressBookProbeMargin, addressBookProbeLimit, AddressBookIDCollisionError
+ Error, addressBookContentHash64), the table builders
(buildAddressBookTable, buildAddressBookTableWithFeeds), the name expansion
(expandBookNameToCIDRs, expandBookNameRecursive), classifyPolicyAddresses,
and the CIDR canonicalization helpers (normalizeAnyInCIDRs, sortV4CIDRs,
sortV6CIDRs, canonicalizeAddressBookContent, dedupSortedStrings). This is
the #1606/#2049/#2514/#3294 deterministic address-book table pipeline.

Pure code-motion, no logic change: byte-identical slices of the
origin/master policies.go. Refs #4421 (codex-173 #4).
